updated feature highlight visualization
